requirements.txt
aioredis==1.3.1
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
Babel==2.9.0
# boto3>=1.20.13
boto3==1.21.28
#botocore==1.23.29
botocore==1.24.28
# cairocffi==0.9.0
CairoSVG==2.5.0
certifi==2020.12.5
cffi==1.14.4
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
cryptography==3.2.1
cssselect2==0.4.1
daphne==2.5.0
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-cors-headers==3.2.1
django-extensions==3.0.3
django-rest-auth==0.9.5
django-smtp-ssl==1.0
django-storages==1.9.1
django-user-activity-log2==0.0.24
djangorestframework==3.11.0
djangorestframework-jwt==1.9.0
docutils==0.15.2
dynamodb-json==1.3
hiredis==2.0.0
html5lib==1.1
hyperlink==20.0.1
idna==2.10
image==1.5.31
imagesize==1.2.0
importlib-metadata==3.4.0
incremental==17.5.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.19.3
packaging==20.7
pandas==1.1.4
paypalrestsdk==1.13.1
Pillow==8.2.0
psycopg2-binary==2.8.6 
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
PyHamcrest==2.0.2
PyJWT==1.4.2
pyOpenSSL==20.0.0
pyparsing==2.4.7
Pyphen==0.10.0
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
qrcode==6.1
razorpay==1.2.0
recommonmark==0.6.0
requests==2.25.0
requests-oauthlib==1.3.0
rinoh-typeface-dejavuserif==0.1.3
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.5.0
s3transfer==0.5.0
sentry-sdk==0.6.8
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.27
sqlparse==0.4.1
tinycss2==1.1.0
twilio==6.4.3a1
Twisted==21.2.0
txaio==20.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
WeasyPrint==51
webencodings==0.5.1
zipp==3.4.0
zope.interface==5.2.0
django-csp==3.7
pymsteams==0.1.16
PyPDF2==1.26.0
mysql-connector-python==8.0.29
openpyxl==3.0.10
pdf2image==1.16.0
poppler-utils==0.1.0
pytest-cov==2.10.0
py==1.9.0
pluggy==0.13.1
# coverage==7.3.0
pytest==6.0.1
django-simple-history==3.0.0